refactoring complete
all code refactored, pending future additions
