+++ src/components/db-components/Systems.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Box, Typography, Stack, Chip } from '@mui/material'
 import { useParams } from 'react-router-dom'
-import { initIPFS, initOrbitDB, getAllDatabases } from '../../database'
+import { initIPFS, initOrbitDB, getAllDatabases } from '../../services/database'
 import { actions, useStateValue } from '../../state'
 
 export function Systems() {
