--- src/database/public/manifest.json
@@ -1,15 +1,0 @@
-{
-  "short_name": "Control Center",
-  "name": "OrbitDB Control Center",
-  "icons": [
-    {
-      "src": "Orbit_round-02.png",
-      "type": "image/png",
-      "sizes": "100x100"
-    }
-  ],
-  "start_url": ".",
-  "display": "standalone",
-  "theme_color": "#000000",
-  "background_color": "#ffffff"
-}
