+++ src/services/database/identities/index.js
@@ -1,0 +1,5 @@
+import NearIdentityProvider from "./NearIdentityProvider";
+
+export{
+    NearIdentityProvider
+}
