+++ src/views/db-views/DBView.jsx
@@ -1,12 +1,12 @@
 import React from 'react'
 import { Outlet } from 'react-router-dom'
 
-import { actions, loadingState, StateProvider } from '../state'
+import { actions, loadingState, StateProvider } from '../../state'
 
-import { Systems } from '../components/db-components/Systems'
-import { Header } from '../components/db-components/Header'
+import { Systems } from '../../components/db-components/Systems'
+import { Header } from '../../components/db-components/Header'
 
-import '../index.css'
+import '../../index.css'
 
 export function DBView() {
 	const initialState = {
